+++ lib/msal-browser/src/telemetry/BrowserPerformanceClient.ts
@@ -19,7 +19,7 @@ export class BrowserPerformanceClient extends PerformanceClient implements IPerf
         this.guidGenerator = new GuidGenerator(this.browserCrypto);
     }
     
-    startPerformanceMeasurement(measureName: string, correlationId: string): IPerformanceMeasurement {
+    startPerformanceMeasuremeant(measureName: string, correlationId: string): IPerformanceMeasurement {
         return new BrowserPerformanceMeasurement(measureName, correlationId);
     }
 
@@ -29,6 +29,12 @@ export class BrowserPerformanceClient extends PerformanceClient implements IPerf
 
     private getPageVisibility(): string | null {
         return document.visibilityState?.toString() || null;
+    }
+
+    supportsBrowserPerformanceNow(): boolean {
+        return typeof window !== "undefined" &&
+            typeof window.performance !== "undefined" &&
+            typeof window.performance.now === "function";
     }
     
     /**
@@ -55,5 +61,64 @@ export class BrowserPerformanceClient extends PerformanceClient implements IPerf
                 });
             }
         };
+    }
+
+    /**
+     * Adds pre-queue time to preQueueTimeByCorrelationId map.
+     * @param {PerformanceEvents} eventName 
+     * @param {?string} correlationId 
+     * @returns 
+     */
+    setPreQueueTime(eventName: PerformanceEvents, correlationId?: string): void {
+        if (!this.supportsBrowserPerformanceNow()) {
+            this.logger.trace(`BrowserPerformanceClient: window performance API not available, unable to set telemetry queue time for ${eventName}`);
+            return;
+        }
+
+        if (!correlationId) {
+            this.logger.trace(`BrowserPerformanceClient: correlationId for ${eventName} not provided, unable to set telemetry queue time`);
+            return;
+        }
+
+        const preQueueTimesByEvents = this.preQueueTimeByCorrelationId.get(correlationId);
+
+        if (preQueueTimesByEvents){
+            preQueueTimesByEvents.set(eventName, window.performance.now());
+            this.preQueueTimeByCorrelationId.set(correlationId, preQueueTimesByEvents);
+        } else {
+            const preQueueTimes = new Map();
+            preQueueTimes.set(eventName, window.performance.now());
+            this.preQueueTimeByCorrelationId.set(correlationId, preQueueTimes);
+        }
+    }
+
+    /**
+     * Calculates and adds queue time measurement for given performance event.
+     * 
+     * @param {PerformanceEvents} name 
+     * @param {?string} correlationId 
+     * @param {?number} preQueueTime 
+     * @returns 
+     */
+    addQueueMeasurement(eventName: PerformanceEvents, correlationId?: string): void {
+        if (!this.supportsBrowserPerformanceNow()) {
+            this.logger.trace(`BrowserPerformanceClient: window performance API not available, unable to add queue measurement for ${eventName}`);
+            return;
+        }
+
+        if (!correlationId) {
+            this.logger.trace(`BrowserPerformanceClient: correlationId for ${eventName} not provided, unable to add queue measurement`);
+            return;
+        }
+
+        const preQueueTime = super.getPreQueueTime(eventName, correlationId);
+        if (!preQueueTime) {
+            return;
+        }
+        
+        const currentTime = window.performance.now();
+        const queueTime = super.calculateQueuedTime(preQueueTime, currentTime);
+
+        return super.addQueueMeasurement(eventName, correlationId, queueTime);
     }
 }