+++ lib/msal-browser/test/telemetry/BrowserPerformanceClient.spec.ts
@@ -40,7 +40,7 @@ describe("BrowserPerformanceClient.ts", () => {
             expect(typeof browserPerfClient.generateId()).toBe("string");
         });
     });
-    describe("startPerformanceMeasurement", () => {
+    describe("startPerformanceMeasuremeant", () => {
         it("calculate performance duration", () => {
             const browserPerfClient = new BrowserPerformanceClient(clientId, authority, logger, name, version, applicationTelemetry, cryptoOptions);
 
@@ -65,5 +65,14 @@ describe("BrowserPerformanceClient.ts", () => {
 
             spy.mockClear();
         });
+    });
+
+    it("supportsBrowserPerformanceNow returns false if window.performance not present", () => {
+        const browserPerfClient = new BrowserPerformanceClient(clientId, authority, logger, name, version, applicationTelemetry, cryptoOptions);
+
+        // @ts-ignore
+        jest.spyOn(window, "performance", "get").mockReturnValue(undefined);
+
+        expect(browserPerfClient.supportsBrowserPerformanceNow()).toBe(false);
     });
 });