+++ lib/msal-common/test/error/InteractionRequiredAuthError.spec.ts
@@ -7,7 +7,11 @@ describe("InteractionRequiredAuthError.ts Class Unit Tests", () => {
     it("InteractionRequiredAuthError object can be created", () => {
         const TEST_ERROR_CODE: string = "test";
         const TEST_ERROR_MSG: string = "This is a test error";
-        const err: InteractionRequiredAuthError = new InteractionRequiredAuthError(TEST_ERROR_CODE, TEST_ERROR_MSG);
+        const TEST_ERROR_TIMESTAMP = "2017-05-01 22:43:20Z";
+        const TEST_ERROR_TRACE_ID = "b72a68c3-0926-4b8e-bc35-3150069c2800";
+        const TEST_ERROR_CORRELATION_ID = "73d656cf-54b1-4eb2-b429-26d8165a52d7";
+        const TEST_ERROR_CLAIMS = '{\"access_token\":{\"polids\":{\"essential\":true,\"values\":[\"9ab03e19-ed42-4168-b6b7-7001fb3e933a\"]}}}';
+        const err: InteractionRequiredAuthError = new InteractionRequiredAuthError(TEST_ERROR_CODE, TEST_ERROR_MSG, "N/A", TEST_ERROR_TIMESTAMP, TEST_ERROR_TRACE_ID, TEST_ERROR_CORRELATION_ID, TEST_ERROR_CLAIMS);
 
         expect(err instanceof InteractionRequiredAuthError).toBe(true);
         expect(err instanceof AuthError).toBe(true);
@@ -16,6 +20,10 @@ describe("InteractionRequiredAuthError.ts Class Unit Tests", () => {
         expect(err.errorMessage).toBe(TEST_ERROR_MSG);
         expect(err.message).toBe(`${TEST_ERROR_CODE}: ${TEST_ERROR_MSG}`);
         expect(err.name).toBe("InteractionRequiredAuthError");
+        expect(err.timestamp).toBe(TEST_ERROR_TIMESTAMP);
+        expect(err.traceId).toBe(TEST_ERROR_TRACE_ID);
+        expect(err.correlationId).toBe(TEST_ERROR_CORRELATION_ID);
+        expect(err.claims).toBe(TEST_ERROR_CLAIMS);
         expect(err.stack?.includes("InteractionRequiredAuthError.spec.ts")).toBe(true);
     });