+++ samples/msal-react-samples/react-router-sample/src/App.js
@@ -1,4 +1,4 @@
-import { Switch, Route, useHistory } from "react-router-dom";
+import { Routes, Route, useNavigate } from "react-router-dom";
 // Material-UI imports
 import Grid from "@mui/material/Grid";
 
@@ -18,45 +18,36 @@ import { ProfileRawContext } from "./pages/ProfileRawContext";
 import { ProfileUseMsalAuthenticationHook } from "./pages/ProfileUseMsalAuthenticationHook";
 
 function App({ pca }) {
-  // The next 3 lines are optional. This is how you configure MSAL to take advantage of the router's navigate functions when MSAL redirects between pages in your app
-  const history = useHistory();
-  const navigationClient = new CustomNavigationClient(history);
-  pca.setNavigationClient(navigationClient);
+    // The next 3 lines are optional. This is how you configure MSAL to take advantage of the router's navigate functions when MSAL redirects between pages in your app
+    const navigate = useNavigate();
+    const navigationClient = new CustomNavigationClient(navigate);
+    pca.setNavigationClient(navigationClient);
 
-  return (
-    <MsalProvider instance={pca}>
-      <PageLayout>
-        <Grid container justifyContent="center">
-          <Pages />
-        </Grid>
-      </PageLayout>
-    </MsalProvider>
-  );
+    return (
+        <MsalProvider instance={pca}>
+            <PageLayout>
+                <Grid container justifyContent="center">
+                    <Pages />
+                </Grid>
+            </PageLayout>
+        </MsalProvider>
+    );
 }
 
 function Pages() {
-  return (
-    <Switch>
-      <Route path="/profile">
-        <Profile />
-      </Route>
-      <Route path="/profileWithMsal">
-        <ProfileWithMsal />
-      </Route>
-      <Route path="/profileRawContext">
-        <ProfileRawContext />
-      </Route>
-      <Route path="/profileUseMsalAuthenticationHook">
-        <ProfileUseMsalAuthenticationHook />
-      </Route>
-      <Route path="/logout">
-          <Logout />
-      </Route>
-      <Route path="/">
-        <Home />
-      </Route>
-    </Switch>
-  )
+    return (
+        <Routes>
+            <Route path="/profile" element={<Profile />} />
+            <Route path="/profileWithMsal" element={<ProfileWithMsal />} />
+            <Route path="/profileRawContext" element={<ProfileRawContext />} />
+            <Route
+                path="/profileUseMsalAuthenticationHook"
+                element={<ProfileUseMsalAuthenticationHook />}
+            />
+            <Route path="/logout" element={<Logout />} />
+            <Route path="/" element={<Home />} />
+        </Routes>
+    );
 }
 
 export default App;